Add files via upload
It is a full stack project i have used html css and bootstrap for front-end and php for backend in this web-app a person can login as a admin or as a user. Admin can post news and they are able to add and delete their users. Users can see the posts of admin.